# src/main/resources/static/api-docs.yml
...
openapi: 3.0.0
info:
  title: "auth-api"
  version: 1.0.0
  description: >-
    auth api.
  license:
    name: Apache-2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
paths:
  /user:
    post:
      tags:
        - user
      description: Create user
      operationId: create
      security:
        - AccessToken: [ ]
        - JwtToken: [ ]
      requestBody:
        required: true
        content:
          application/vnd.kaiqkt_auth_api_user_v1+json:
            schema:
              $ref: '#/components/schemas/UserRequestV1'
      responses:
        200:
          description: User created
          content:
            application/vnd.kaiqkt_auth_api_user_v1+json:
              schema:
                $ref: '#/components/schemas/UserResponseV1'
        400:
          description: Email already exists, invalid password/email, or invalid first/last name
          content:
            application/vnd.kaiqkt_auth_api_user_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
    get:
      tags:
        - user
      description: Get user by token
      operationId: findByToken
      security:
        - JwtToken: [ ]
      responses:
        200:
          description: User found
          content:
            application/vnd.kaiqkt_auth_api_user_v1+json:
              schema:
                $ref: '#/components/schemas/UserResponseV1'
        404:
          description: User not found
          content:
            application/vnd.kaiqkt_auth_api_user_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
...
  /user/{user_id}:
    get:
      tags:
        - user
      description: Get user by id
      operationId: findById
      security:
        - AccessToken: [ ]
        - JwtToken: [ ]
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
          required: true
      responses:
        200:
          description: User found
          content:
            application/vnd.kaiqkt_auth_api_user_v1+json:
              schema:
                $ref: '#/components/schemas/UserResponseV1'
        404:
          description: User not found
          content:
            application/vnd.kaiqkt_auth_api_user_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
  /user/{user_id}/status/{status}:
    patch:
      tags:
        - user
      description: Update user status
      operationId: updateStatus
      security:
        - JwtToken: [ ]
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
          required: true
        - in: path
          name: status
          schema:
            type: string
          required: true
      responses:
        204:
          description: User status updated
  /user/{user_id}/role/{role_id}:
    patch:
      tags:
        - user
      description: Update user role
      operationId: updateRole
      security:
        - JwtToken: [ ]
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
          required: true
        - in: path
          name: role_id
          schema:
            type: string
          required: true
      responses:
        204:
          description: User roles updated
    delete:
      tags:
        - user
      description: Remove user role
      operationId: removeRole
      security:
        - JwtToken: [ ]
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
          required: true
        - in: path
          name: role_id
          schema:
            type: string
          required: true
      responses:
        204:
          description: User roles updated
...
  /credential:
    get:
      tags:
        - credential
      description: Get credential information by token
      operationId: findByToken
      security:
        - JwtToken: [ ]
      responses:
        200:
          description: Credential found
          content:
            application/vnd.kaiqkt_auth_api_credential_v1+json:
              schema:
                $ref: '#/components/schemas/CredentialResponseV1'
        404:
          description: Credential not found
          content:
            application/vnd.kaiqkt_auth_api_credential_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
    patch:
      tags:
        - credential
      description: Change credential
      operationId: change
      security:
        - JwtToken: [ ]
      requestBody:
        required: true
        content:
          application/vnd.kaiqkt_auth_api_credential_v1+json:
            schema:
              $ref: '#/components/schemas/CredentialRequestV1'
      responses:
        204:
          description: Credential updated
        404:
          description: Credential not found
          content:
            application/vnd.kaiqkt_auth_api_credential_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
...
components:
  schemas:
...
    UserRequestV1:
      type: object
      properties:
        first_name:
          type: string
          example: "John"
          x-field-extra-annotation: '@NotNull(message="first name should be not null") @Size(max = 255, message = "first name must be between 2 and 14 characters")'
        last_name:
          type: string
          example: "Doe"
          x-field-extra-annotation: '@NotNull(message="last name should be not null") @Size(min= 1, max = 255, message = "last name must be between 2 and 14 characters")'
        email:
          type: string
          example: "email@auth-api.com"
          x-field-extra-annotation: '@NotNull(message="email name should be not null") @Email(message = "email must be a valid email address")'
        password:
          type: string
          example: "password"
          x-field-extra-annotation: '@NotNull(message="password should be not null") @Pattern(regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%^&+=]).{8,}$",message = "Password must contain at least one lowercase letter, one uppercase letter, one digit, one special character, and be at least 8 characters long.")'
      required:
        - first_name
        - last_name
        - email
        - password
    UserResponseV1:
      type: object
      properties:
        id:
          type: string
          example: "01J2N1XP9M1Q4YK2YPMQE78V1D"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        email:
          type: string
          example: "email@auth-api.com"
        status:
          type: string
          example: "ACTIVE"
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleResponseV1'
        created_at:
          type: string
          example: "2021-08-01T00:00:00Z"
        updated_at:
          type: string
          example: "2021-08-01T00:00:00Z"
      required:
        - id
        - first_name
        - last_name
        - email
        - roles
        - status
        - created_at
        - updated_at
    RoleResponseV1:
      type: object
      properties:
        id:
          type: string
          example: "01J2N1XP9M1Q4YK2YPMQE78V1D"
        name:
          type: string
          example: "ADMIN"
        created_at:
          type: string
          example: "2021-08-01T00:00:00Z"
      required:
        - id
        - name
        - created_at
...
    CredentialResponseV1:
      type: object
      properties:
        id:
          type: string
          example: "01J2N1XP9M1Q4YK2YPMQE78V1D"
        user_id:
          type: string
          example: "01J2N1XP9M1Q4YK2YPMQE78V1D"
        created_at:
          type: string
          example: "2021-08-01T00:00:00Z"
        updated_at:
          type: string
          example: "2021-08-01T00:00:00Z"
      required:
        - id
        - user_id
        - created_at
        - updated_at
    CredentialRequestV1:
      type: object
      properties:
        password:
          type: string
          example: "password"
          x-field-extra-annotation: '@NotNull(message="password should be not null")'
        new_password:
          type: string
          example: "password"
          x-field-extra-annotation: '@NotNull(message="password should be not null") @Pattern(regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%^&+=]).{8,}$",message = "Password must contain at least one lowercase letter, one uppercase letter, one digit, one special character, and be at least 8 characters long.")'
...
    ErrorV1:
      type: object
      properties:
        type:
          type: string
          example: "INVALID_PAYLOAD"
        message:
          type: object
          example: "Invalid property 'name' with value 'null'"
      required:
        - type
        - message
  securitySchemes:
    JwtToken:
      type: http
      scheme: bearer
    AccessToken:
      type: apiKey
      name: Authorization
      in: header